+++ src/app/components/references/index.js
@@ -43,6 +43,10 @@ angular
                       return 'Macros';
                 } else if (type == 'exposure') {
                       return 'Exposures';
+                } else if (type == 'metric') {
+                      return 'Metrics';
+                } else if (type == 'operation') {
+                        return 'Operations';
                 } else {
                     return 'Nodes';
                 }