+++ src/app/components/table_details/table_details.js
@@ -148,8 +148,9 @@ angular
                 var get_type = _.property(['metadata', 'type'])
                 var rel_type = get_type(nv);
 
-                scope.meta = nv.meta || null;
-
+                var sources_meta = nv.hasOwnProperty('sources') ? nv.sources[0] != undefined ? nv.sources[0].source_meta : null : null;
+                scope.meta = nv.meta || sources_meta;
+                
                 scope.details = getBaseStats(nv);
                 scope.extended = getExtendedStats(nv.stats);