+++ src/app/services/graph.service.js
@@ -186,6 +186,12 @@ angular
                     }
                 },
                 {
+                    selector: 'node[resource_type="metric"]',
+                    style: {
+                        'background-color': '#ff5688',
+                    }
+                },
+                {
                     selector: 'node[selected=1]',
                     style: {
                         'background-color': '#bd6bb6',
@@ -329,9 +335,10 @@ angular
 
 
         _.each(_.filter(service.manifest.nodes, function(node) {
-            var is_graph_type = _.includes(['model', 'seed', 'source', 'snapshot', 'analysis', 'exposure'], node.resource_type);
-            var is_data_test = node.resource_type == 'test' && _.includes(node.tags, 'data');
-            return is_graph_type || is_data_test;
+            // operation needs to be a graph type so that the parent/child mpa can be resolved even though we won't be displaying it
+            var is_graph_type = _.includes(['model', 'seed', 'source', 'snapshot', 'analysis', 'exposure', 'metric', 'operation'], node.resource_type);
+            var is_singular_test = node.resource_type == 'test' && !node.hasOwnProperty('test_metadata');
+            return is_graph_type || is_singular_test;
         }), function(node) {
             var node_obj = {
                 group: "nodes",
@@ -352,7 +359,7 @@ angular
 
                 if (!_.includes(['model', 'source', 'seed', 'snapshot'], parent_node.resource_type)) {
                     return;
-                } else if (child_node.resource_type == 'test' && _.includes(child_node.tags, 'schema')) {
+                } else if (child_node.resource_type == 'test' && child_node.hasOwnProperty('test_metadata')) {
                     return;
                 }