+++ src/app/services/selector_matcher.test.js
@@ -80,7 +80,7 @@ const elements = [
         data: {
             id: 7,
             resource_type: 'test',
-            tags: ['data'],
+            tags: [],
             name: 'test_unique_page_views_id',
             original_file_path: 'tests/test_unique.yml',
             package_name: 'my_package',
@@ -99,7 +99,7 @@ test("Test FQN Matching glob", () => {
     expect(
         matcher.isFQNMatch(
             ['my_package', 'dir', 'model'],
-            ['my_package', '*']
+            'my_package.*'
         )
     ).toStrictEqual(true)
 })
@@ -108,7 +108,16 @@ test("Test FQN Matching all parts", () => {
     expect(
         matcher.isFQNMatch(
             ['my_package', 'dir', 'model'],
-            ['my_package', 'dir', 'model']
+            'my_package.dir.model'
+        )
+    ).toStrictEqual(true)
+})
+
+test("Test FQN Matching all parts with dots", () => {
+    expect(
+        matcher.isFQNMatch(
+            ['my_package', 'dir', 'ns1.ns2.model'],
+            'my_package.dir.ns1.ns2.model'
         )
     ).toStrictEqual(true)
 })
@@ -117,7 +126,7 @@ test("Test FQN Matching bare package", () => {
     expect(
         matcher.isFQNMatch(
             ['my_package', 'dir', 'model'],
-            ['my_package']
+            'my_package'
         )
     ).toStrictEqual(true)
 })
@@ -126,7 +135,7 @@ test("Test FQN Matching bare path", () => {
     expect(
         matcher.isFQNMatch(
             ['my_package', 'dir', 'model'],
-            ['my_package', 'dir']
+            'my_package.dir'
         )
     ).toStrictEqual(true)
 })
@@ -135,7 +144,7 @@ test("Test FQN Matching bare path glob", () => {
     expect(
         matcher.isFQNMatch(
             ['my_package', 'dir', 'model'],
-            ['my_package', 'dir', '*']
+            'my_package.dir.*'
         )
     ).toStrictEqual(true)
 })
@@ -144,7 +153,7 @@ test("Test FQN Matching glob all", () => {
     expect(
         matcher.isFQNMatch(
             ['my_package', 'dir', 'model'],
-            ['*']
+            '*'
         )
     ).toStrictEqual(true)
 })
@@ -153,7 +162,16 @@ test("Test FQN Matching direct model name", () => {
     expect(
         matcher.isFQNMatch(
             ['my_package', 'dir', 'model'],
-            ['model']
+            'model'
+        )
+    ).toStrictEqual(true)
+})
+
+test("Test FQN Matching direct model name with dots", () => {
+    expect(
+        matcher.isFQNMatch(
+            ['my_package', 'dir', 'ns1.ns2.model'],
+            'ns1.ns2.model'
         )
     ).toStrictEqual(true)
 })
@@ -162,7 +180,7 @@ test("Test FQN Matching non match package", () => {
     expect(
         matcher.isFQNMatch(
             ['my_package', 'dir', 'model'],
-            ['other_package']
+            'other_package'
         )
     ).toStrictEqual(false)
 })
@@ -171,7 +189,7 @@ test("Test FQN Matching non match path", () => {
     expect(
         matcher.isFQNMatch(
             ['my_package', 'dir', 'model'],
-            ['my_package', 'other_dir']
+            'my_package.other_dir'
         )
     ).toStrictEqual(false)
 })
@@ -428,9 +446,21 @@ test("Test getting nodes by test type", () => {
     )
 
     expect(
+        matchByTestType('singular')
+    ).toStrictEqual(
+        [7]
+    )
+
+    expect(
         matchByTestType('schema')
     ).toStrictEqual(
         []
+    )
+
+    expect(
+        matchByTestType('generic')
+    ).toStrictEqual(
+        [7]
     )
 
     expect(