+++ src/app/sources/index.js
@@ -29,6 +29,9 @@ angular
             if (sources.length == 0) {
                 return;
             }
+            
+            // sort sources by sources.name
+            sources.sort((a,b) => a.name.localeCompare(b.name));
 
             var source = sources[0];